Webapp/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
dj-database-url==2.1.0
Django==4.2.8
django-rest-knox==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
heroku==0.1.4
idna==3.6
packaging==23.2
Pillow==10.1.0
pycparser==2.21
pyscopg2==66.0.2
python-dateutil==1.5
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.8.0
Unidecode==1.3.7
urllib3==2.1.0
whitenoise==6.6.0
